feat (post): move post form to separate component
